// src/app/model/packages/provider/hotel-room-package.ts
export class HotelRoomPackage {
    id: string;
    hotelId: string;
    roomType: string;
    roomDescription: string;
    customerInformationPack: string;
    adultCount: number;
    childCount: number;
    pricePerNight: number;
    currency: string;
    discount: number;
    breakfast: boolean;
}